Switch to BigDecimal, rename INCOME -> DEPOSIT, minor changes

+++ app/src/main/java/com/example/bittrack/data/mappers/TransactionMapper.kt
@@ -3,11 +3,12 @@ package com.example.bittrack.data.mappers
 import com.example.bittrack.data.local.model.TransactionEntity
 import com.example.bittrack.domain.models.Transaction
 import com.example.bittrack.core.util.DateUtils
+import java.math.BigDecimal
 
 fun Transaction.toEntity(): TransactionEntity {
     return TransactionEntity(
         id = id,
-        amount = amount,
+        amount = amount.toDouble(),
         type = type,
         category = category,
         timestamp = DateUtils.localDateTimeToEpochMillis(timestamp)
@@ -17,7 +18,7 @@ fun Transaction.toEntity(): TransactionEntity {
 fun TransactionEntity.toDomain(): Transaction {
     return Transaction(
         id = id,
-        amount = amount,
+        amount = BigDecimal.valueOf(amount),
         type = type,
         category = category,
         timestamp = DateUtils.epochMillisToLocalDateTime(timestamp)
